added console

+++ src/main/java/ch/uzh/ifi/seal/soprafs17/service/UserService.java
@@ -48,8 +48,10 @@ public class UserService {
 
     public void deleteUser(Long id) {
         User user = userRepository.findById(id); //TODO check if user exists
-        userRepository.delete(id);
-        log.debug("Deleted User: {}", user);
+        if (user != null) {
+            userRepository.delete(id);
+            log.debug("Deleted User: {}", user);
+        }
     }
